resolve float problem in migration

+++ database/migrations/2020_05_30_101329_create_purchased_items_table.php
@@ -13,7 +13,7 @@ class CreatePurchasedItemsTable extends Migration
             $table->id();
             $table->integer('quotation_id');
             $table->integer('request_item_id');
-            $table->float('purchased_price');
+            $table->integer('purchased_price');
             $table->date('purchased_date');
             $table->string('purchased_name');
             $table->string('buyer');
